Extract pasted-list spacing into PastedNodesNormalizer
Move the list-spacing normalization out of lexical_helper.js into a
dedicated PastedNodesNormalizer class, mirroring PastedContentFormatter.
Consolidate the formatting and normalization paste tests into a single
pasted_text_formatting.test.js with two test groups.
