+++ src/editor/contents.js
@@ -18,7 +18,8 @@ import { $createActionTextAttachmentUploadNode, ActionTextAttachmentUploadNode }
 import { $getNearestBlockElementAncestorOrThrow } from "@lexical/utils"
 import NodeInserter from "./contents/node_inserter"
 import PastedContentFormatter from "./contents/pasted_content_formatter"
-import { $consecutiveSiblingGroups, $expandSelectionToLineBreaksAndSplitAtEdges, $isShadowRoot, $normalizeListSpacing, $splitSelectedParagraphsAtInnerLineBreaks } from "../helpers/lexical_helper"
+import PastedNodesNormalizer from "./contents/pasted_nodes_normalizer"
+import { $consecutiveSiblingGroups, $expandSelectionToLineBreaksAndSplitAtEdges, $isShadowRoot, $splitSelectedParagraphsAtInnerLineBreaks } from "../helpers/lexical_helper"
 
 export default class Contents {
   constructor(editorElement) {
@@ -42,7 +43,7 @@ export default class Contents {
       if ($hasUpdateTag(PASTE_TAG)) this.#formatPastedDOM(doc)
 
       let nodes = this.editorElement.$generateNodesFromDOM(doc)
-      if ($hasUpdateTag(PASTE_TAG)) nodes = $normalizeListSpacing(nodes)
+      if ($hasUpdateTag(PASTE_TAG)) nodes = new PastedNodesNormalizer(nodes).normalize()
 
       if (!$hasUpdateTag(PASTE_TAG) || !this.#dispatchPastedNodesCommand(nodes)) {
         this.#insertUploadNodes(nodes) || this.insertAtCursor(...nodes)
